requirements.txt
mtcnn
tensorflow
keras
opencv-python
numpy
pillow
tk
